improve: leaderboard structure, nft img added
